+++ contrib/activity/file/src/main/scala/com/krux/hyperion/contrib/activity/file/RepartitionFile.scala
@@ -3,8 +3,7 @@ package com.krux.hyperion.contrib.activity.file
 import java.io.{File, FilenameFilter}
 import java.nio.file._
 
-import com.krux.hyperion.contrib.activity.file.enumeration.CompressionFormat
-import com.krux.hyperion.contrib.activity.file.enumeration.CompressionFormat.CompressionFormat
+import com.krux.hyperion.contrib.activity.file.CompressionFormat.CompressionFormat
 import scopt.OptionParser
 
 object RepartitionFile {